+++ torch/csrc/jit/tensorexpr/loopnest.cpp
@@ -6,7 +6,10 @@
 #include <unordered_set>
 #include <vector>
 
+#include <c10/util/Logging.h>
 #include <torch/csrc/jit/tensorexpr/eval.h>
+#include <torch/csrc/jit/tensorexpr/expr.h>
+#include <torch/csrc/jit/tensorexpr/ir.h>
 #include <torch/csrc/jit/tensorexpr/ir_mutator.h>
 #include <torch/csrc/jit/tensorexpr/ir_printer.h>
 #include <torch/csrc/jit/tensorexpr/tensor.h>
